+++ CHANGES.md
@@ -5,6 +5,7 @@ History
 
 * Add `pyppeteer-install` command
 * Add `autoClose` option to `launch` fucntion
+* Add `loop` option to `launch` function (experimental)
 * Continue to find WS URL while process is alive
 
 ## Version 0.0.20 (2018-08-11)